refactor(aggbuffer): change the size type from int to int64_t

+++ arl/am/am_agg.hpp
@@ -256,7 +256,7 @@ Future<std::invoke_result_t<Fn, Args...>> run_lpc(rank_t context, Fn&& fn, Args&
 void flush_amagg_buffer() {
   for (int ii = 0; ii < proc::rank_n(); ++ii) {
     int i = (ii + local::rank_me()) % proc::rank_n();
-    std::vector<std::pair<char*, int>> results = amagg_agg_buffer_p[i].flush();
+    std::vector<std::pair<char*, int64_t>> results = amagg_agg_buffer_p[i].flush();
     for (auto result: results) {
       if (std::get<0>(result) != nullptr) {
         if (std::get<1>(result) != 0) {
@@ -334,7 +334,7 @@ Future<std::invoke_result_t<Fn, Args...>> rpc_agg(rank_t remote_worker, Fn&& fn,
 //  printf("send meta: %ld, %ld, %d\n", meta.fn_p, meta.type_wrapper_p, meta.target_local_rank);
 //  printf("sizeof(payload): %lu\n", sizeof(Payload));
 
-  std::pair<char*, int> result = amagg_internal::amagg_agg_buffer_p[remote_proc].push(meta, std::move(payload));
+  std::pair<char*, int64_t> result = amagg_internal::amagg_agg_buffer_p[remote_proc].push(meta, std::move(payload));
   if (std::get<0>(result) != nullptr) {
     if (std::get<1>(result) != 0) {
       gex_AM_RequestMedium0(backend::tm, remote_proc, amagg_internal::hidx_gex_amagg_reqhandler,
